complete design details for blogs

+++ src/services/designServices.js
@@ -1,7 +1,14 @@
 import axios from "api/aixos"
+import { toast } from "react-toastify";
+
+const isAdmin = sessionStorage.getItem('isAdmin');
 
 const createDetail = async (createdData) => {
-    return await axios.post(`/api/design-details`, createdData);
+    if (isAdmin) {
+        return await axios.post(`/api/design-details`, createdData);
+    } else {
+        toast.error('Tài khoản này không cho phép tạo dữ liệu');
+    }
 }
 
 const getAllDetail = async () => {
@@ -12,18 +19,31 @@ const getDetailById = async (detailId) => {
     return await axios.get(`/api/design-details/${detailId}`);
 }
 
+const getDetailByBlogId = async (blogId) => {
+    return await axios.get(`/api/design-details/blogs/${blogId}`);
+}
+
 const updateDetail = async (detailId, updatedData) => {
-    return await axios.put(`/api/design-details/${detailId}`, updatedData);
+    if (isAdmin) {
+        return await axios.put(`/api/design-details/${detailId}`, updatedData);
+    } else {
+        toast.error('Tài khoản này không cho phép cập nhật dữ liệu');
+    }
 }
 
 const deleteDetail = async (detailId) => {
-    return await axios.delete(`/api/design-details/${detailId}`);
+    if (isAdmin) {
+        return await axios.delete(`/api/design-details/${detailId}`);
+    } else {
+        toast.error('Tài khoản này không cho phép xóa dữ liệu');
+    }
 }
 
 export default {
     createDetail,
     getAllDetail,
     getDetailById,
+    getDetailByBlogId,
     updateDetail,
     deleteDetail
 }
